+++ src/main/java/com/clearing/json/CMTrade.java
@@ -3,9 +3,9 @@ package com.clearing.json;
 public class CMTrade {
 	private String es;
 	private int qty;
-	private float price;
-	private float transactionAmt;
-	public CMTrade(String es, int qty, float price, float transactionAmt) {
+	private double price;
+	private double transactionAmt;
+	public CMTrade(String es, int qty, double price, double transactionAmt) {
 		super();
 		this.es = es;
 		this.qty = qty;
@@ -24,16 +24,16 @@ public class CMTrade {
 	public void setQty(int qty) {
 		this.qty = qty;
 	}
-	public float getPrice() {
+	public double getPrice() {
 		return price;
 	}
-	public void setPrice(float price) {
+	public void setPrice(double price) {
 		this.price = price;
 	}
-	public float getTransactionAmt() {
+	public double getTransactionAmt() {
 		return transactionAmt;
 	}
-	public void setTransactionAmt(float transactionAmt) {
+	public void setTransactionAmt(double transactionAmt) {
 		this.transactionAmt = transactionAmt;
 	}
