+++ src/test/java/com/clearing/TradeServiceImplStub.java
@@ -44,9 +44,9 @@ public class TradeServiceImplStub implements TradeService {
 	}
 
 	@Override
-	public Pair<HashMap<Integer, Float>, HashMap<Integer, HashMap<Integer, Integer>>> hashMapifyTrades() {
+	public Pair<HashMap<Integer, Double>, HashMap<Integer, HashMap<Integer, Integer>>> hashMapifyTrades() {
 		List<TradeEntity> tradesList = tradeRepository.findAll();
-		HashMap<Integer, Float> transactionAmountHashMap = new HashMap<Integer, Float>();
+		HashMap<Integer, Double> transactionAmountHashMap = new HashMap<Integer, Double>();
 		HashMap<Integer, HashMap<Integer, Integer>> quantityHashMap = new HashMap<Integer, HashMap<Integer, Integer>>();
 
 		for (TradeEntity trade : tradesList) {
@@ -57,13 +57,13 @@ public class TradeServiceImplStub implements TradeService {
 			if (!transactionAmountHashMap.containsKey(buyerCMId)) {
 				transactionAmountHashMap.put(buyerCMId, -1 * trade.getTransactionAmount());
 			} else {
-				Float updatedTransactionAmount = transactionAmountHashMap.get(buyerCMId) - trade.getTransactionAmount();
+				Double updatedTransactionAmount = transactionAmountHashMap.get(buyerCMId) - trade.getTransactionAmount();
 				transactionAmountHashMap.put(buyerCMId, updatedTransactionAmount);
 			}
 			if (!transactionAmountHashMap.containsKey(sellerCMId)) {
 				transactionAmountHashMap.put(sellerCMId, trade.getTransactionAmount());
 			} else {
-				Float updatedTransactionAmount = transactionAmountHashMap.get(sellerCMId)
+				Double updatedTransactionAmount = transactionAmountHashMap.get(sellerCMId)
 						+ trade.getTransactionAmount();
 				transactionAmountHashMap.put(sellerCMId, updatedTransactionAmount);
 			}
@@ -101,7 +101,7 @@ public class TradeServiceImplStub implements TradeService {
 
 		}
 
-		return new Pair<HashMap<Integer, Float>, HashMap<Integer, HashMap<Integer, Integer>>>(transactionAmountHashMap,
+		return new Pair<HashMap<Integer, Double>, HashMap<Integer, HashMap<Integer, Integer>>>(transactionAmountHashMap,
 				quantityHashMap);
 	}
